address some feedback
